+++ packages/lucia-sveltekit/src/ambient.d.ts
@@ -1,8 +1,11 @@
 declare namespace Lucia {
-    export interface UserData {} 
+    export interface UserAttributesSchema {} 
+    export interface User {}
 }
 
 /// <reference types="@sveltejs/kit" />
 declare namespace App {
-	export interface Locals {}
+	export interface Locals {
+        getSession: () => import("./types.js").Session | null
+    }
 }