+++ packages/lucia-sveltekit/src/types.ts
@@ -1,14 +1,12 @@
 import type { LuciaError } from "./error.js";
 
-export type User = Required<Lucia.UserData> & {
-    userId: string;
-};
+export type User = Lucia.User extends never ? { userId: string } : Lucia.User;
 
 export type Session = {
-    sessionId: string
+    sessionId: string;
     userId: string;
     expires: number;
-}
+};
 
 export type Env = "DEV" | "PROD";
 export type Error = typeof LuciaError;